feat(Prefabs): add default rule to ignore special grab items
The StringInList rule is now being used as a default rule applied
to the DistanceGrabber so the distance grabber by default ignores
any Interactable that has a String Observable List with the element
`specialgrab` in the list.

This can be used to automatically ignore certain Interactables from
being able to be position grabbed.