feat(Facade): add pre/post/cancel events for grab
The Facade now has three new events when the pointer grabs an item.
It will emit a BeforeGrabbed event when the grab is initiated and
a AfterGrabbed when the grab is finished. If the grab is canceled
during transition then a GrabCanceled event is emitted.

Also fixed an issue with the pointer cursor not looking correct in
VR. It now just uses an unlit transparent color shader instead of
the screen shader and it uses a mesh instead of billboards so it
looks better in the headset. It also uses a normal rotator so the
cursor is always flat on the surface it is hitting.