fix(prefab): remove race condition on clear of grabbing interactable
There was a race condition where the grabbing interactable was getting
cleared before the grab took place, this was more prominent if the
grab precognition on the interactor was higher. This has been fixed
by putting the clear logic after a end of frame yield emitter so
it ensures everything is done before the clear is executed.