feat(prefab): replace interactable grabber with base component
The InteractableGrabber component now exists in the Interactables
package so this version is now being used as there is already a
dependency on this package anyway.

The InteractableGrabber has been deleted from this package now to
prevent any clashes. This is technically a breaking change for anyone
who would be using this component.