fix(prefabs): ensure pointer is re-enabled if interactable disabled
If the Interactable is disabled or destroyed after the pointer grabber
has grabbed it, then the relevant logic to re-enable the pointer
grabber wasn't being run and therefore the pointer grabber was no
longer working.

This fix checks to see if the grabbed interactable is still available
in the scene and if its not then it forcefully calls the enable pointer
logic to ensure it is re-enabled correctly.