fix(Configurator): reenable pointer on interactable disabled on ungrab
This fixes an issue where if the grabbed interactable is disabled when
it has been ungrabbed but the untouched event does not occur in time
then the pointer would not re-enable. This now waits a frame and checks
to see if the ungrabbed object is still active in the scene and if not
then it re-enables the pointer logic as expected.