fix(Configurator): honour the interactable disabled states
The Interactable can have the touch and grab states disabled meaning
the interactable is not touchable/grabbable. However, the pointer
interactor was still attempting to touch and grab the interactable
even if the Touch/Grab was disabled.

This fix attempts to short circuit the logic by checking the
interactable touch and grab state in the appropriate point in the
logic.